# api/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: NAS Control Plane API
  version: 0.1.0
  description: Phase 0 的 API 契约。能力信息只能由 ncp-agent 的受控接口提供。
paths:
...
  /api/v1/system/dns/preview:
    post:
      operationId: previewDNSChange
      summary: 预览 DNS 修改，不立即写入系统
      tags: [system]
      security: [{RootSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNSChangeRequest'
      responses:
        '200':
          description: DNS 修改预览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSChangePreview'
        '409':
          description: 当前 DNS 后端只读或无法预览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/system/dns/confirm:
    post:
      operationId: confirmDNSChange
      summary: 确认并应用 DNS 修改
      tags: [system]
      security: [{RootSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNSChangeConfirmation'
      responses:
        '200':
          description: DNS 修改结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSChangeResult'
        '409':
          description: 确认失败或预览已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/system/dns/rollback:
    post:
      operationId: rollbackDNSChange
      summary: 回滚已应用的 DNS 修改
      tags: [system]
      security: [{RootSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNSRollbackRequest'
      responses:
        '200':
          description: DNS 回滚结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSChangeResult'
        '409':
          description: 回滚失败或变更已不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/system/public-egress/capability:
    get:
      operationId: getPublicEgressCapability
      summary: 获取公网出口检测能力
      description: 只读取部署配置，不会自动访问外部检测端点。
      tags: [system]
      security: [{RootSession: []}]
      responses:
        '200':
          description: 公网出口检测能力
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEgressCapability'
  /api/v1/system/public-egress/detect:
    post:
      operationId: detectPublicEgress
      summary: 用户显式触发公网出口检测
      tags: [system]
      security: [{RootSession: []}]
      responses:
        '200':
          description: 公网出口地址及可选地区信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEgressResult'
        '503':
          description: 检测端点未配置、不可达或返回无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEgressResult'
...
  /api/v1/proxy/mihomo/invoke:
    post:
      operationId: invokeMihomo
      summary: 调用受 allowlist 保护的 Mihomo 操作
      tags: [proxy]
      security: [{RootSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MihomoInvokeRequest'
      responses:
        '200':
          description: Mihomo 控制器响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MihomoInvokeResult'
        '400':
          description: 操作不在 allowlist 或参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 控制器不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MihomoInvokeRequest:
      type: object
      required: [operation]
      properties:
        operation: { type: string, enum: [version, health, traffic, connections, proxies, select-proxy] }
        group: { type: string, maxLength: 128 }
        proxy: { type: string, maxLength: 128 }
    MihomoInvokeResult:
      type: object
      required: [operation, statusCode, data]
      properties:
        operation: { type: string }
        statusCode: { type: integer, minimum: 100, maximum: 599 }
        data: {}
...
    DNSChangeRequest:
      type: object
      required: [nameservers]
      properties:
        interface: { type: string, maxLength: 128 }
        connectionId: { type: string, maxLength: 256 }
        nameservers:
          type: array
          minItems: 1
          maxItems: 6
          items: { type: string, format: ipv4 }
        searchDomains:
          type: array
          maxItems: 8
          items: { type: string, maxLength: 253 }
    DNSState:
      type: object
      required: [interface, connectionId, nameservers, searchDomains]
      properties:
        interface: { type: string }
        connectionId: { type: string }
        nameservers: { type: array, items: { type: string } }
        searchDomains: { type: array, items: { type: string } }
    DNSChangePreview:
      type: object
      required: [previewId, backend, before, after, requiresConfirm, rollbackAvailable, expiresAt, errorCode]
      properties:
        previewId: { type: string }
        backend: { type: string }
        before: { $ref: '#/components/schemas/DNSState' }
        after: { $ref: '#/components/schemas/DNSState' }
        requiresConfirm: { type: boolean }
        rollbackAvailable: { type: boolean }
        expiresAt: { type: string, format: date-time }
        errorCode: { type: string }
    DNSChangeConfirmation:
      type: object
      required: [previewId, confirmed]
      properties:
        previewId: { type: string, minLength: 1 }
        confirmed: { type: boolean, const: true }
    DNSRollbackRequest:
      type: object
      required: [changeId]
      properties:
        changeId: { type: string, minLength: 1 }
    DNSChangeResult:
      type: object
      required: [changeId, backend, applied, rollbackAvailable, appliedAt, errorCode]
      properties:
        changeId: { type: string }
        backend: { type: string }
        applied: { type: boolean }
        rollbackAvailable: { type: boolean }
        appliedAt: { type: string, format: date-time }
        errorCode: { type: string }
    PublicEgressCapability:
      type: object
      required: [configured, status, endpoint, requiresUserAction, detectionSource, errorCode]
      properties:
        configured: { type: boolean }
        status: { type: string }
        endpoint: { type: string, description: 已脱敏的检测端点 }
        requiresUserAction: { type: boolean }
        detectionSource: { type: string }
        errorCode: { type: string }
    PublicEgressResult:
      type: object
      required: [status, address, checkedAt, detectionSource, errorCode]
      properties:
        status: { type: string }
        address: { type: string }
        country: { type: string }
        region: { type: string }
        isp: { type: string }
        asn: { type: string }
        checkedAt: { type: string, format: date-time }
        detectionSource: { type: string }
        errorCode: { type: string }
...
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
...
  securitySchemes:
    RootSession:
      type: apiKey
      in: cookie
      name: ncp_session